[Android] fix NDK extract location
The Android NDK is being extracted with an extra intermediate directory:

  $BinaryCache\android-ndk-r26b\android-ndk-r26b

The Android build fails because it expects the NDK contents to live
immediately under the directory:

  $BinaryCache\android-ndk-r26b

The extra android-ndk-r26b sub-dir comes from the NDK zip archive
itself, which we have no control over. To work-around, allow callers
of Extract-ZipFile to control whether or not it should create the
extraction sub-dir. Override the default behavior for NDK extraction
only.