CMake: Improving missing compiler error message
Right now, if you have the new parser enabled and try to configure
without an existing Swift compiler, the configuration fails due to a
call to 'get_filename_component' with the wrong number of arguments.
This is is because the `CMAKE_Swift_COMPILER` is empty and doesn't
expand to anything here:

```
get_filename_component(swift_bin_dir ${CMAKE_Swift_COMPILER} DIRECTORY)
```

The current error is not clear about why it failed though. Improving the
error message so that other people can look at it and see that it's
because the compiler is missing.